+++ casino/src/main/java/game/Bingo.java
@@ -1,7 +1,7 @@
 package game;
 
 public class Bingo extends GameAbstract {
-    public Bingo(int maxPlayers, int minBet, double cashOutMult, double winChance, int gameTime){
+    public Bingo(int maxPlayers, double minBet, double cashOutMult, double winChance, int gameTime){
         super("Bingo", maxPlayers, minBet, cashOutMult, winChance, gameTime);
     }
 }
