+++ casino/src/main/java/game/Slotmachine.java
@@ -1,7 +1,7 @@
 package game;
 
 public class Slotmachine extends GameAbstract {
-    public Slotmachine(int maxPlayers, int minBet, double cashOutMult, double winChance, int gameTime){
+    public Slotmachine(int maxPlayers, double minBet, double cashOutMult, double winChance, int gameTime){
         super("Slotmachine", maxPlayers, minBet, cashOutMult, winChance, gameTime);
     }
 }
